Danit Uziel test
Thank you for this test. I have made  the popup as close to the mockup. Made a prefab and animated some movements. As for the art replacement tool, i do not know this feature and will take me some time to research it. so that's it for now.
Thank you :)